% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa_cdc.R
\docType{data}
\name{acdc_fonts}
\alias{acdc_fonts}
\title{Africa CDC fonts}
\format{
An object of class \code{list} of length 4.
}
\usage{
acdc_fonts
}
\description{
Africa CDC fonts
}
\examples{
acdc_fonts

}
\keyword{datasets}
