% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_acdc.R
\name{acdc_palettes}
\alias{acdc_palettes}
\alias{acdc_brewer_palettes}
\title{Africa CDC palettes}
\usage{
acdc_palettes

acdc_brewer_palettes
}
\description{
Africa CDC palettes
}
\examples{
acdc_palettes

acdc_brewer_palettes

}
