% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleta_brewer.R
\name{paleta_create_sequential}
\alias{paleta_create_sequential}
\alias{paleta_create_divergent}
\alias{paleta_create_qualitative}
\alias{paleta_create_brewer}
\title{Create new palettes based on organisational palettes}
\usage{
paleta_create_sequential(n, org = c("acdc", "nhs"), name)

paleta_create_divergent(n, name, org)

paleta_create_qualitative(n, name, org)

paleta_create_brewer(
  n,
  name,
  org,
  type = c("sequential", "divergent", "qualitative")
)
}
\arguments{
\item{n}{Number of colours desired/required. Organisational palettes should
have at least 3 colours and up to 9 colours maximum. All colour schemes are
derived from an organisation's brand/style guidelines.}

\item{org}{Name of organisation. Currently supports only \emph{"acdc"} for the
Africa CDC colour palettes.}

\item{name}{Name of the organisational palette to use}

\item{type}{A character value for type of palette to use. Can be either
\emph{"sequential"}, \emph{"divergent"}, or \emph{"qualitative"}.}
}
\value{
A character vector of desired/required colours with length
equivalent to \code{n}
}
\description{
These functions apply a similar approach used and demonstrated by
\href{https://colorbrewer2.org}{ColorBrewer} and has been patterned after the
syntax of the \code{RColorBrewer} package
}
\examples{
paleta_create_sequential(n = 5, org = "acdc", name = "blues")
paleta_create_divergent(n = 5, org = "acdc", name = "rdylgn")

}
