% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_colours.R
\name{get_colour}
\alias{get_colour}
\alias{get_colours}
\title{Get named colours vector}
\usage{
get_colour(
  pattern = NULL,
  model = c("hex", "rgb", "cmyk", "pantone"),
  named = FALSE
)

get_colours(
  pattern = NULL,
  model = c("hex", "rgb", "cmyk", "pantone"),
  named = FALSE
)
}
\arguments{
\item{pattern}{Optional. A character value or vector to use as a search term.
Default is NULL in which case all the Oxford colours are returned.}

\item{model}{A character vector of colour model. Can be "rgb", "cmyk", "hex",
or "pantone". Default is "hex".}

\item{named}{Logical. Should the output be a named character value or
vector? Default is FALSE.}
}
\value{
A character value or vector of colour/s as per `model` specification.
  If `named` is TRUE, returns a named character value or vector.
}
\description{
Get named colours vector
}
\examples{
get_colours()
get_colours(model = "rgb")
get_colours(pattern = "orange")
get_colours(pattern = c("orange", "brown"), named = TRUE)

}
