% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gnc.R
\name{gnc_green}
\alias{gnc_green}
\alias{gnc_grey}
\alias{gnc_sage}
\alias{gnc_dark_green}
\alias{gnc_lime}
\alias{gnc_aqua_green}
\alias{gnc_cyan}
\alias{gnc_blue}
\alias{gnc_orange}
\alias{gnc_moss_green}
\alias{gnc_gold}
\title{GNC colours}
\usage{
gnc_green

gnc_grey

gnc_sage

gnc_dark_green

gnc_lime

gnc_aqua_green

gnc_cyan

gnc_blue

gnc_orange

gnc_moss_green

gnc_gold
}
\description{
GNC colours
}
\examples{
gnc_green
gnc_grey
gnc_sage

}
