% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gnc.R
\name{gnc_fonts}
\alias{gnc_fonts}
\alias{set_gnc_font}
\title{GNC fonts}
\usage{
gnc_fonts

set_gnc_font(alt = paleta_fonts$paleta_noto)
}
\arguments{
\item{alt}{A character value for font family to use if all of the GNC fonts
are not available in the system.}
}
\value{
A character value for font family to use as GNC font.
}
\description{
The function will search the system for availability of any of the GNC fonts
in hierarchical order starting with \emph{Arial}, then \emph{Calibri Regular}, then
\emph{Source Sans Pro}, and then finally \emph{Arial}. If none of these are found in
the system, the function will return \emph{Noto Sans} by default or the user can
set which font to use as alternative by specifying \code{alt}.
}
\examples{
gnc_fonts

set_gnc_font()

}
