% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/motif.R
\docType{package}
\name{motif}
\alias{motif-package}
\alias{motif}
\title{Collection of Palette, Theme, and Theme Components}
\description{
A collection of palettes, themes, and theme components based on branding
guidelines of various international non-governmental organisations and
United Nations units.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://katilingban.io/motif/}
}

}
\author{
Ernest Guevarra (ernest@guevarra.io)
}
\keyword{internal}
