% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_motif.R
\name{set_motif_font}
\alias{set_motif_font}
\title{Set motif font based on what is available in the system}
\usage{
set_motif_font()
}
\value{
A character value for font family to use as motif font.
}
\description{
The function will search the system for availability of any of the motif
fonts in heirarchical order starting with \emph{Arial Narrow}, then \emph{Noto Sans},
and then \emph{Roboto}.
}
\examples{
set_motif_font()

}
