% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_motif.R
\docType{data}
\name{motif_fonts}
\alias{motif_fonts}
\title{Base package fonts}
\format{
An object of class \code{list} of length 3.
}
\usage{
motif_fonts
}
\description{
Base package fonts
}
\examples{
motif_fonts

}
\keyword{datasets}
