% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nhs.R
\name{nhs_palettes}
\alias{nhs_palettes}
\alias{nhs_brewer_palettes}
\title{NHS palettes}
\usage{
nhs_palettes

nhs_brewer_palettes
}
\description{
NHS palettes
}
\examples{
nhs_palettes

nhs_brewer_palettes

}
