% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores_dubois.R
\name{obtener_paleta_completa}
\alias{obtener_paleta_completa}
\title{Obtener la paleta de colores discreta}
\usage{
obtener_paleta_completa()
}
\value{
Un vector con los colores definidos.
}
\description{
Esta función devuelve la paleta de colores predefinidos.
}
