% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores_dubois.R
\name{obtener_paleta_conam}
\alias{obtener_paleta_conam}
\title{Obtener la paleta de colores continua en amarillos}
\usage{
obtener_paleta_conam()
}
\value{
Un vector con los colores definidos.
}
\description{
Esta función devuelve la paleta alternativa de colores predefinidos.
}
