% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores_dubois.R
\name{obtener_paleta_congr}
\alias{obtener_paleta_congr}
\title{Obtener la paleta de colores continua en grises}
\usage{
obtener_paleta_congr()
}
\value{
Un vector con los colores definidos.
}
\description{
Esta función devuelve la paleta alternativa de colores predefinidos.
}
