% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleta_brewer.R
\docType{data}
\name{paleta_brewer_types}
\alias{paleta_brewer_types}
\title{Palette types}
\format{
An object of class \code{list} of length 3.
}
\usage{
paleta_brewer_types
}
\description{
Palette types
}
\keyword{internal}
