% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleta_brewer.R
\name{paleta_check_colour}
\alias{paleta_check_colour}
\title{Check if a colour palette name is from a specified organisation}
\usage{
paleta_check_colour(name, org)
}
\description{
Check if a colour palette name is from a specified organisation
}
\keyword{internal}
