% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleta_brewer.R
\name{paleta_check_type}
\alias{paleta_check_type}
\title{Check if a colour palette is divergent, sequential, or qualitative}
\usage{
paleta_check_type(name, pal_type = c("sequential", "divergent", "qualitative"))
}
\description{
Check if a colour palette is divergent, sequential, or qualitative
}
\keyword{internal}
