% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_paleta.R
\name{set_paleta_font}
\alias{set_paleta_font}
\title{Set paleta font based on what is available in the system}
\usage{
set_paleta_font()
}
\value{
A character value for font family to use as paleta font.
}
\description{
The function will search the system for availability of any of the paleta
fonts in hierarchical order starting with \emph{Roboto Condensed}, then
\emph{Noto Sans}, and then \emph{Roboto}.
}
\examples{
set_paleta_font()

}
