% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_paleta.R
\docType{data}
\name{paleta_fonts}
\alias{paleta_fonts}
\title{Base package fonts}
\format{
An object of class \code{list} of length 3.
}
\usage{
paleta_fonts
}
\description{
Base package fonts
}
\examples{
paleta_fonts

}
\keyword{datasets}
