% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.palette}
\alias{print.palette}
\title{Print function for palettes}
\usage{
\method{print}{palette}(x, ...)
}
\arguments{
\item{x}{An object of class \code{palette}}

\item{...}{Additional arguments to print}
}
\description{
Print function for palettes
}
