% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores_dubois.R
\name{scale_continuous_am}
\alias{scale_continuous_am}
\title{Paleta de colores en amarillo para variables continuas}
\usage{
scale_continuous_am()
}
\value{
Un scale_color_gradientn para usar con ggplot2.
}
\description{
Esta función devuelve un \code{scale_color_gradientn()} con los tonos en amarillo
}
