% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores_dubois.R
\name{scale_continuous_gr}
\alias{scale_continuous_gr}
\title{Paleta de colores en gris para variables continuas}
\usage{
scale_continuous_gr()
}
\value{
Un scale_color_gradientn para usar con ggplot2.
}
\description{
Esta función devuelve un \code{scale_color_gradientn()} con los tonos en gris
}
