% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colores_dubois.R
\name{scale_dubois_discreta}
\alias{scale_dubois_discreta}
\title{Paleta de colores nueva para variables discretas}
\usage{
scale_dubois_discreta()
}
\value{
Un scale_color_manual para usar con ggplot2.
}
\description{
Esta función devuelve un \code{scale_color_manual()} con los colores definidos en la paleta nueva.
}
