% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{shade_colour}
\alias{shade_colour}
\alias{shade_colours_}
\alias{shade_colours}
\title{Get shade of colours}
\usage{
shade_colour(hex, p)

shade_colours_(hex, p)

shade_colours(hex, p, label = FALSE)
}
\arguments{
\item{hex}{A character value or vector of character of values for hex code
of colour/s to shade.}

\item{p}{Range from 0 to 1 for proportion to shade the colour/s with.}

\item{label}{Logical. Should the output/s be labelled? Default is FALSE.}
}
\value{
A character value or vector of character values of hex code/s
shaded to the desired proportion.
}
\description{
Get shade of colours
}
\examples{
shade_colour(acdc_green, p = 0.2)
shade_colours(acdc_palettes$acdc_secondary, p = 0.4)

}
