% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nhs.R
\name{theme_nhs}
\alias{theme_nhs}
\title{A \link[ggplot2:ggplot2]{ggplot2} theme using NHS fonts, colours, and palettes}
\usage{
theme_nhs(
  base_family = set_nhs_font(),
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_colour = nhs_blue,
  subtitle_family = base_family,
  subtitle_colour = nhs_dark_grey,
  caption_colour = nhs_dark_grey,
  axis_title_colour = nhs_dark_grey,
  legend_title_colour = nhs_dark_grey,
  legend_text_colour = nhs_dark_grey,
  grid_col = nhs_mid_grey,
  grid = TRUE,
  axis_col = nhs_mid_grey,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{Base font family using NHS fonts. Default is set
by what NHS font is available in the system via \code{\link[=set_nhs_font]{set_nhs_font()}}.
If none of the NHS fonts are available, the default becomes
\emph{Noto Sans}.}

\item{base_size}{Base font size. Default is 11.5.}

\item{plot_title_family}{Font family to use for the plot title. Default is
\code{base_family}.}

\item{plot_title_colour}{Colour of the plot title text. Default
is \link{nhs_blue}.}

\item{subtitle_family}{Font family to use for the plot subtitle. Default is
`base_family``.}

\item{subtitle_colour}{Colour of the subtitle text. Default is \link{nhs_mid_grey}.}

\item{caption_colour}{Colour of the caption text. Default is \link{nhs_mid_grey}.}

\item{axis_title_colour}{Colour of the axis title text. Default is
\link{nhs_mid_grey}.}

\item{legend_title_colour}{Colour of the legend title text. Default is NULL.}

\item{legend_text_colour}{Colour of the legend text. Default is NULL.}

\item{grid_col}{Grid colour. Default to \link{nhs_pale_grey}.}

\item{grid}{Panel grid. Either \code{TRUE}, \code{FALSE}, or a combination of
\code{X} (major x grid), \code{x} (minor x grid), \code{Y} (major y grid), and/or
\code{y} (minor y grid). Default is TRUE.}

\item{axis_col}{Axis colours. Default to \link{nhs_pale_grey}.}

\item{axis}{Add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}". Default is FALSE.}

\item{ticks}{Logical. Should ticks be added? Default is FALSE.}
}
\value{
A \link[ggplot2:ggplot2]{ggplot2} theme.
}
\description{
These are wrappers for \code{\link[=theme_paleta]{theme_paleta()}} that use colours and fonts from the
NHS visual identity guidelines.
}
\section{Colours}{

The NHS theme is based on the colours from the \link{nhs_palettes}. The
primary palette consists of two colours: \code{nhs_palettes$nhs_primary}. The
secondary palette consists of nineteen colours:.
}

\section{Fonts}{

The NHS theme uses two fonts as prescribed by the NHS visual
identity guidelines. These fonts (in hierarchical order of preference) are
\emph{Frutiger} and \emph{Arial}. Any or all of these fonts should be available in
the user's system for them to be used in the theme. If none of these fonts
are available in the user's system, a freely downloadable alternative called
\emph{Noto Sans} is the default fallback font and can be obtained from
\href{https://fonts.google.com/}{Google Fonts}.
}

\examples{
\dontrun{
  ggplot(
    data = mtcars,
    mapping = aes(
      x = factor(vs, levels = c(0, 1), labels = c("v-shaped", "straight")),
      fill = factor(cyl))
  ) +
  geom_bar() +
  scale_fill_manual(
    name = "Cylinders",
    values = nhs_palettes$nhs_support_greens
  ) +
  labs(
    title = "Engine shape by number of cylinders",
    subtitle = "An example plot for this package",
    x = "Engine Shape",
    y = "Counts"
   ) +
   theme_nhs()
}

}
