% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_paleta.R
\name{theme_paleta}
\alias{theme_paleta}
\title{A generic/base \link[ggplot2:ggplot2]{ggplot2} theme for the paleta package}
\usage{
theme_paleta(
  base_family = set_paleta_font(),
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 16,
  plot_title_face = "bold",
  plot_title_colour = NULL,
  plot_title_margin = 10,
  subtitle_family = base_family,
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_colour = NULL,
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = base_family,
  caption_size = 9,
  caption_face = "italic",
  caption_colour = NULL,
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = subtitle_family,
  axis_title_size = 9,
  axis_title_colour = NULL,
  axis_title_face = "plain",
  axis_title_just = "rt",
  legend_title_family = subtitle_family,
  legend_title_colour = NULL,
  legend_text_family = subtitle_family,
  legend_text_colour = NULL,
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  plot_background_fill = NULL,
  grid_col = NULL,
  grid = TRUE,
  axis_col = NULL,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{Base font family using Africa CDC fonts. Default is set
by what Africa CDC font is available in the system via \code{\link[=set_paleta_font]{set_paleta_font()}}.}

\item{base_size}{Base font size. Default is 11.5.}

\item{plot_title_family}{Font family to use for the plot title. Default is
\code{base_family}.}

\item{plot_title_size}{Plot title text size in pts. Default is 16.}

\item{plot_title_face}{Font face ("plain", "italic", "bold", "bold.italic")
for plot title. Default is "bold".}

\item{plot_title_colour}{Colour of the plot title text. Default
is \code{acdc_green}.}

\item{plot_title_margin}{Margin at the bottom of the plot title. Default
set at 10.}

\item{subtitle_family}{Font family to use for the plot subtitle. Default is
\code{base_family}.}

\item{subtitle_size}{Plot subtitle text size in pts. Default is 12.}

\item{subtitle_face}{Font face ("plain", "italic", "bold", "bold.italic")
for plot subtitle. Default is "plain".}

\item{subtitle_colour}{Colour of the subtitle text. Default is NULL to use
default \link[ggplot2:ggplot2]{ggplot2} colour.}

\item{subtitle_margin}{Margin at the bottom of the plot subtitle. Default
set at 15.}

\item{strip_text_family}{Font family to use for the facet label. Default is
\code{base_family}.}

\item{strip_text_size}{Facet label text size in pts. Default is 12.}

\item{strip_text_face}{Font face ("plain", "italic", "bold", "bold.italic")
for facet label. Default is "plain".}

\item{caption_family}{Font family to use for the caption text. Default is
\code{base_family}.}

\item{caption_size}{Caption text size in pts. Default is 9.}

\item{caption_face}{Font face ("plain", "italic", "bold", "bold.italic") for
caption text. Default is "plain".}

\item{caption_colour}{Colour of the caption text. Default is NULL.}

\item{caption_margin}{Margin at the top of the plot caption text. Default is
set at 10.}

\item{axis_text_size}{Axis text size in pts. Default is \code{base_size}.}

\item{axis_title_family}{Font family to use for the axis title. Default is
\code{subtitle_family}.}

\item{axis_title_size}{Axis title text size in pts. Default is 9.}

\item{axis_title_colour}{Colour of the axis title text. Default is NULL to
use default \link[ggplot2:ggplot2]{ggplot2} colour.}

\item{axis_title_face}{Font face ("plain", "italic", "bold", "bold.italic")
for axis title. Default is "plain".}

\item{axis_title_just}{Axis title font justification, one of
"bl" (bottom-left), "m" (middle), "rt" (right-top). Default is "rt".}

\item{legend_title_family}{Font family to use for the legend title. Default
is \code{subtitle_family}.}

\item{legend_title_colour}{Colour of the legend title text. Default is NULL.}

\item{legend_text_family}{Font family to use for the legend text. Default
is \code{subtitle_family}.}

\item{legend_text_colour}{Colour of the legend text. Default is NULL.}

\item{plot_margin}{Plot margins (specify with \code{\link[ggplot2:margin]{ggplot2::margin()}}).}

\item{plot_background_fill}{Fill colour for the plot background. Default is
NULL.}

\item{grid_col}{Grid colour. Default to NULL.}

\item{grid}{Panel grid. Either \code{TRUE}, \code{FALSE}, or a combination of
\code{X} (major x grid), \code{x} (minor x grid), \code{Y} (major y grid), and/or
\code{y} (minor y grid). Default is TRUE.}

\item{axis_col}{Axis colours. Default to NULL.}

\item{axis}{Add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}". Default is FALSE.}

\item{ticks}{Logical. Should ticks be added? Default is FALSE.}
}
\value{
A \link[ggplot2:ggplot2]{ggplot2} theme.
}
\description{
This generic/base \link[ggplot2:ggplot2]{ggplot2} theme is the template from which all other
themes in the \link{paleta} package is built on. This theme is inspired by
Bob Rudis' \href{https://github.com/hrbrmstr/hrbrthemes}{hrbrthemes} package
drawing heavily on its typography-centric focus.
}
\details{
This function uses either of three fonts - \emph{Roboto Condensed}, \emph{Noto Sans},
or \emph{Roboto} (in this specific order) - depending on what is available in the
user's system. \emph{Roboto Condensed}, \emph{Noto Sans}, and \emph{Roboto} - all Google
fonts - are freely downloadable and easily installed on any system.
If any of these fonts is not available in your system, download and install
either/all \emph{Roboto Condensed}, \emph{Noto Sans} and/or \emph{Roboto} from
\href{https://fonts.google.com/}{Google Fonts}.
}
