% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_gnc.R, R/theme_unicef.R
\name{theme_gnc}
\alias{theme_gnc}
\alias{theme_unicef}
\title{A \link[ggplot2:ggplot2]{ggplot2} theme using GNC fonts, colours, and palettes}
\usage{
theme_gnc(
  base_family = set_gnc_font(),
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_colour = gnc_grey,
  subtitle_family = base_family,
  subtitle_colour = gnc_grey,
  caption_colour = gnc_grey,
  axis_title_colour = gnc_grey,
  legend_title_colour = gnc_grey,
  legend_text_colour = gnc_grey,
  grid_col = gnc_grey,
  grid = TRUE,
  axis_col = gnc_grey,
  axis = FALSE,
  ticks = FALSE
)

theme_unicef(
  base_family = set_unicef_font(),
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_colour = unicef_black,
  subtitle_family = base_family,
  subtitle_colour = unicef_cool_grey,
  caption_colour = unicef_cool_grey,
  axis_title_colour = unicef_cool_grey,
  legend_title_colour = unicef_cool_grey,
  legend_text_colour = unicef_cool_grey,
  grid_col = unicef_warm_grey,
  grid = TRUE,
  axis_col = unicef_warm_grey,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{Base font family using UNICEF fonts. Default is set
by what UNICEF font is available in the system via \code{\link[=set_unicef_font]{set_unicef_font()}}.
If none of the UNICEF fonts are available, the default becomes
\emph{Noto Sans}.}

\item{base_size}{Base font size. Default is 11.5.}

\item{plot_title_family}{Font family to use for the plot title. Default is
\code{base_family}.}

\item{plot_title_colour}{Colour of the plot title text. Default
is \link{unicef_black}.}

\item{subtitle_family}{Font family to use for the plot subtitle. Default is
\code{base_family}.}

\item{subtitle_colour}{Colour of the subtitle text. Default is
\link{unicef_cool_grey}.}

\item{caption_colour}{Colour of the caption text. Default is
\link{unicef_cool_grey}.}

\item{axis_title_colour}{Colour of the axis title text. Default is
\link{unicef_cool_grey}.}

\item{legend_title_colour}{Colour of the legend title text. Default is NULL.}

\item{legend_text_colour}{Colour of the legend text. Default is NULL.}

\item{grid_col}{Grid colour. Default to \link{unicef_warm_grey}.}

\item{grid}{Panel grid. Either \code{TRUE}, \code{FALSE}, or a combination of
\code{X} (major x grid), \code{x} (minor x grid), \code{Y} (major y grid), and/or
\code{y} (minor y grid). Default is TRUE.}

\item{axis_col}{Axis colours. Default to \link{unicef_warm_grey}.}

\item{axis}{Add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}". Default is FALSE.}

\item{ticks}{Logical. Should ticks be added? Default is FALSE.}
}
\value{
A \link[ggplot2:ggplot2]{ggplot2} theme.

A \link[ggplot2:ggplot2]{ggplot2} theme.
}
\description{
These are wrappers for \code{\link[=theme_paleta]{theme_paleta()}} that use colours and fonts from the
GNC visual identity guidelines.

These are wrappers for \code{\link[=theme_paleta]{theme_paleta()}} that use colours and fonts from the
UNICEF visual identity guidelines.
}
\section{Colours}{

The GNC theme is based on the colours from the \link{gnc_palettes}. The
primary palette consists of one colour: \code{gnc_palettes$gnc_primary}.
The secondary palette consists of ten colours:
\code{gnc_palettes$gnc_secondary}.


The UNICEF theme is based on the colours from the \link{unicef_palettes}. The
primary palette consists of one colour: \code{unicef_palettes$unicef_primary}.
The secondary palette consists of eleven colours:
\code{unicef_palettes$unicef_secondary}.
}

\section{Fonts}{

The GNC theme uses four fonts as prescribed by the GNC visual identity
guidelines. These fonts (in hierarchical order of preference) are
\emph{Arial}, \emph{Calibri}, \emph{Source Sans Pro}, and \emph{Bebas Neue}. Any or all of these
fonts should be available in the user's system for them to be used in the
theme. If none of these fonts are available in the user's system, a freely
downloadable alternative called \emph{Noto Sans} is the default fallback font and
can be obtained from \href{https://fonts.google.com/}{Google Fonts}.


The UNICEF theme uses four fonts as prescribed by the UNICEF visual identity
guidelines. These fonts (in hierarchical order of preference) are
\emph{Univers LT Pro}, \emph{Arial}, \emph{Roboto}, and \emph{Aleo}. Any or all of these fonts
should be available in the user's system for them to be used in the theme.
If none of these fonts are available in the user's system, a freely
downloadable alternative called \emph{Noto Sans} is the default fallback font and
can be obtained from \href{https://fonts.google.com/}{Google Fonts}.
}

\examples{
\dontrun{
  ggplot(
    data = mtcars,
    mapping = aes(
      x = factor(vs, levels = c(0, 1), labels = c("v-shaped", "straight")),
      fill = factor(cyl))
  ) +
  geom_bar() +
  scale_fill_manual(
    name = "Cylinders",
    values = gnc_palettes$gnc_secondary
  ) +
  labs(
    title = "Engine shape by number of cylinders",
    subtitle = "An example plot for this package",
    x = "Engine Shape",
    y = "Counts"
   ) +
   theme_unicef()
}

\dontrun{
  ggplot(
    data = mtcars,
    mapping = aes(
      x = factor(vs, levels = c(0, 1), labels = c("v-shaped", "straight")),
      fill = factor(cyl))
  ) +
  geom_bar() +
  scale_fill_manual(
    name = "Cylinders",
    values = unicef_palettes$unicef_secondary
  ) +
  labs(
    title = "Engine shape by number of cylinders",
    subtitle = "An example plot for this package",
    x = "Engine Shape",
    y = "Counts"
   ) +
   theme_unicef()
}

}
