% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_unicef.R
\name{unicef_fonts}
\alias{unicef_fonts}
\alias{set_unicef_font}
\title{UNICEF fonts}
\usage{
unicef_fonts

set_unicef_font(alt = paleta_fonts$paleta_noto)
}
\arguments{
\item{alt}{A character value for font family to use if all of the UNICEF
fonts are not available in the system.}
}
\value{
A character value for font family to use as UNICEF font.
}
\description{
The function will search the system for availability of any of the UNICEF
fonts in hierarchical order starting with \emph{Andes}, and then \emph{Arial}. If
none of these are found in the system, the function will return \emph{Noto Sans}
by default or the user can set which font to use as alternative by specifying
\code{alt}.
}
\examples{
unicef_fonts

set_unicef_font()

}
