% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\docType{data}
\name{wb_palettes}
\alias{wb_palettes}
\title{World Bank palettes}
\format{
An object of class \code{list} of length 18.
}
\usage{
wb_palettes
}
\description{
World Bank palettes
}
\examples{
wb_palettes

}
\keyword{datasets}
