% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa_cdc.R
\docType{data}
\name{acdc_green}
\alias{acdc_green}
\alias{acdc_gold}
\alias{acdc_black}
\alias{acdc_brown}
\alias{acdc_yellow}
\alias{acdc_orange}
\alias{acdc_red}
\alias{acdc_dark_green}
\title{Africa CDC colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
acdc_green

acdc_gold

acdc_black

acdc_brown

acdc_yellow

acdc_orange

acdc_red

acdc_dark_green
}
\description{
Africa CDC colours
}
\examples{
acdc_green
acdc_gold
acdc_brown

}
\keyword{datasets}
