% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa_cdc.R
\name{set_acdc_font}
\alias{set_acdc_font}
\title{Set ACDC font to use based on what is available from the system}
\usage{
set_acdc_font(alt = motif_fonts$motif_noto)
}
\arguments{
\item{alt}{A character value for font family to use if all of the Africa
CDC fonts are not available in the system.}
}
\value{
A character value for font family to use as Africa CDC font.
}
\description{
The function will search the system for availability of any of the Africa
CDC fonts in heirarchical order starting with *Acumin Pro*, then
*Brandon Text*, then *Calibri*, and then finally *Arial*. If none of these
are found in the system, the function will return *Noto Sans* by default or
the user can set which font to use as alternative by specifying `alt`.
}
\examples{
set_acdc_font()

}
