% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_acdc.R
\docType{data}
\name{acdc_fonts}
\alias{acdc_fonts}
\alias{set_acdc_font}
\title{Africa CDC fonts}
\format{
An object of class \code{list} of length 3.
}
\usage{
acdc_fonts

set_acdc_font(alt = paleta_fonts$paleta_noto)
}
\arguments{
\item{alt}{A character value for font family to use if all of the Africa
CDC fonts are not available in the system.}
}
\value{
A character value for font family to use as Africa CDC font.
}
\description{
The function will search the system for availability of any of the Africa
CDC fonts in hierarchical order starting with \emph{Univers}, then
\emph{Din Next Arabic}, and then finally \emph{Arial}. If none of these
are found in the system, the function will return \emph{Noto Sans} by default or
the user can set which font to use as alternative by specifying \code{alt}.
}
\examples{
acdc_fonts

set_acdc_font()

}
\keyword{datasets}
