% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_acdc.R
\docType{data}
\name{acdc_palettes}
\alias{acdc_palettes}
\title{Africa CDC palettes}
\format{
An object of class \code{list} of length 2.
}
\usage{
acdc_palettes
}
\description{
Africa CDC palettes
}
\examples{
acdc_palettes

}
\keyword{datasets}
