% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{motif_colours}
\alias{motif_colours}
\title{Colours based on visual identity guidelines of various organisations}
\format{
A data frame with 5 fields

| **Variable** | **Description** |
| :--- | :--- |
| *name* | Organisation colour name |
| *code* | Organisation colour code |
| *rgb* | Three integers for the red, green, blue components of the RGB colour model |
| *cmyk* | Four integers for the cyan, magenta, yellow, and black components of the CMYK colour model |
| *hex* | Hexadecimal codes for corresponding colour |
| *pantone* | Pantone colour name |
}
\usage{
motif_colours
}
\description{
Colours based on visual identity guidelines of various organisations
}
\examples{
motif_colours


}
\keyword{datasets}
