% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nhs.R
\docType{data}
\name{nhs_blue}
\alias{nhs_blue}
\alias{nhs_white}
\alias{nhs_dark_blue}
\alias{nhs_bright_blue}
\alias{nhs_light_blue}
\alias{nhs_aqua_blue}
\alias{nhs_black}
\alias{nhs_dark_grey}
\alias{nhs_mid_grey}
\alias{nhs_pale_grey}
\alias{nhs_dark_green}
\alias{nhs_green}
\alias{nhs_light_green}
\alias{nhs_aqua_green}
\alias{nhs_purple}
\alias{nhs_dark_pink}
\alias{nhs_pink}
\alias{nhs_dark_red}
\alias{nhs_orange}
\alias{nhs_warm_yellow}
\alias{nhs_yellow}
\title{NHS colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
nhs_blue

nhs_white

nhs_dark_blue

nhs_bright_blue

nhs_light_blue

nhs_aqua_blue

nhs_black

nhs_dark_grey

nhs_mid_grey

nhs_pale_grey

nhs_dark_green

nhs_green

nhs_light_green

nhs_aqua_green

nhs_purple

nhs_dark_pink

nhs_pink

nhs_dark_red

nhs_orange

nhs_warm_yellow

nhs_yellow
}
\description{
NHS colours
}
\examples{
nhs_blue
nhs_dark_blue

}
\keyword{datasets}
