% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_nhs.R
\docType{data}
\name{nhs_fonts}
\alias{nhs_fonts}
\alias{set_nhs_font}
\title{NHS fonts}
\format{
An object of class \code{list} of length 2.
}
\usage{
nhs_fonts

set_nhs_font(alt = paleta_fonts$paleta_noto)
}
\arguments{
\item{alt}{A character value for font family to use if all of the NHS
fonts are not available in the system.}
}
\value{
A character value for font family to use as NHS font.
}
\description{
The function will search the system for availability of any of the NHS
fonts in heirarchical order starting with \emph{Andes}, and then \emph{Arial}. If
none of these are found in the system, the function will return \emph{Noto Sans}
by default or the user can set which font to use as alternative by specifying
\code{alt}.
}
\examples{
nhs_fonts

set_nhs_font()

}
\keyword{datasets}
