% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paleta.R
\docType{package}
\name{paleta}
\alias{paleta-package}
\alias{paleta}
\title{Collection of Palette, Theme, and Theme Components}
\description{
A collection of palettes, themes, and theme components based on branding
guidelines of various international non-governmental organisations and
United Nations units.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/katilingban/paleta,http://katilingban.io/paleta/}
}

}
\author{
Ernest Guevarra (ernest@guevarra.io)
}
\keyword{internal}
