% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/africa_cdc.R
\name{theme_acdc_light}
\alias{theme_acdc_light}
\title{A [ggplot2] theme using Africa CDC fonts, colours, and palettes}
\usage{
theme_acdc_light(
  base_family = set_acdc_font(),
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_size = 18,
  plot_title_face = "bold",
  plot_title_colour = acdc_green,
  plot_title_margin = 10,
  subtitle_family = base_family,
  subtitle_size = 12,
  subtitle_face = "plain",
  subtitle_colour = acdc_gold,
  subtitle_margin = 15,
  strip_text_family = base_family,
  strip_text_size = 12,
  strip_text_face = "plain",
  caption_family = base_family,
  caption_size = 9,
  caption_face = "italic",
  caption_colour = acdc_green,
  caption_margin = 10,
  axis_text_size = base_size,
  axis_title_family = subtitle_family,
  axis_title_size = 9,
  axis_title_colour = acdc_gold,
  axis_title_face = "plain",
  axis_title_just = "rt",
  plot_margin = ggplot2::margin(30, 30, 30, 30),
  grid_col = acdc_gold,
  grid = TRUE,
  axis_col = acdc_gold,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{Base font family. Default is set by what is available in
the sytem via `set_acdc_font()`.}

\item{base_size}{Base font size. Default is 11.5}

\item{plot_title_family, plot_title_face, plot_title_size, plot_title_margin, plot_title_colour}{plot title family, face, size and margi}

\item{subtitle_family, subtitle_face, subtitle_size, subtitle_colour}{plot subtitle family, face and size}

\item{subtitle_margin}{plot subtitle margin bottom (single numeric value)}

\item{strip_text_family, strip_text_face, strip_text_size}{facet label font family, face and size}

\item{caption_family, caption_face, caption_size, caption_margin, caption_colour}{plot caption family, face, size and margin}

\item{axis_text_size}{font size of axis text}

\item{axis_title_family, axis_title_face, axis_title_size, axis_title_colour}{axis title font family, face and size}

\item{axis_title_just}{axis title font justification, one of `[blmcrt]`}

\item{plot_margin}{plot margin (specify with `ggplot2::margin()`)}

\item{grid_col, axis_col}{grid & axis colors; both default to `#cccccc`}

\item{grid}{panel grid (`TRUE`, `FALSE`, or a combination of `X`, `x`, `Y`, `y`)}

\item{axis}{add x or y axes? `TRUE`, `FALSE`, "`xy`"}

\item{ticks}{ticks if `TRUE` add ticks}
}
\value{
A [ggplot2] theme.
}
\description{
A [ggplot2] theme using Africa CDC fonts, colours, and palettes
}
\examples{
\dontrun{
  ggplot(
    data = mtcars,
    mapping = aes(
      x = factor(vs, levels = c(0, 1), labels = c("v-shaped", "straight")),
      fill = factor(cyl))
  ) +
  geom_bar() +
  scale_fill_manual(
    name = "Cylinders",
    values = acdc_palettes$acdc_secondary
  ) +
  labs(
    title = "Engine shape by number of cylinders",
    subtitle = "An example plot for this package",
    x = "Engine Shape",
    y = "Counts"
   ) +
   theme_acdc_light()
}

}
