% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_acdc.R
\name{theme_acdc_light}
\alias{theme_acdc_light}
\alias{theme_acdc_dark}
\title{A \link{ggplot2} theme using Africa CDC fonts, colours, and palettes}
\usage{
theme_acdc_light(
  base_family = set_acdc_font(),
  base_size = 11.5,
  plot_title_family = base_family,
  plot_title_colour = acdc_green,
  subtitle_family = base_family,
  subtitle_colour = acdc_gold,
  caption_colour = acdc_gold,
  axis_title_colour = acdc_gold,
  legend_title_colour = acdc_gold,
  legend_text_colour = acdc_gold,
  grid_col = acdc_gold,
  grid = TRUE,
  axis_col = acdc_gold,
  axis = FALSE,
  ticks = FALSE
)

theme_acdc_dark(
  base_family = set_acdc_font(),
  plot_title_family = base_family,
  plot_title_colour = acdc_green,
  subtitle_family = base_family,
  subtitle_colour = acdc_black,
  caption_colour = acdc_black,
  axis_title_colour = acdc_black,
  legend_title_colour = acdc_black,
  legend_text_colour = acdc_black,
  plot_background_fill = tint_colour(acdc_gold, 0.2),
  grid_col = acdc_green,
  grid = TRUE,
  axis_col = acdc_green,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{Base font family using Africa CDC fonts. Default is set
by what Africa CDC font is available in the system via \code{set_acdc_font()}.
If none of the Africa CDC fonts are available, the default becomes Noto
Sans.}

\item{base_size}{Base font size. Default is 11.5.}

\item{plot_title_family}{Font family to use for the plot title. Default is
\code{base_family}.}

\item{plot_title_colour}{Colour of the plot title text. Default
is \code{acdc_green}.}

\item{subtitle_family}{Font family to use for the plot subtitle. Default is
\code{base_family}.}

\item{subtitle_colour}{Colour of the subtitle text. Default is \code{acdc_gold}.}

\item{caption_colour}{Colour of the caption text. Default is \code{acdc_gold}.}

\item{axis_title_colour}{Colour of the axis title text. Default is
\code{acdc_gold}.}

\item{legend_title_colour}{Colour of the legend title text. Default is NULL.}

\item{legend_text_colour}{Colour of the legend text. Default is NULL.}

\item{grid_col}{Grid colour. Default to \code{acdc_gold}.}

\item{grid}{Panel grid. Either \code{TRUE}, \code{FALSE}, or a combination of
\code{X} (major x grid), \code{x} (minor x grid), \code{Y} (major y grid), and/or
\code{y} (minor y grid). Default is TRUE.}

\item{axis_col}{Axis colours. Default to \code{acdc_gold}.}

\item{axis}{Add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}". Default is FALSE.}

\item{ticks}{Logical. Should ticks be added? Default is FALSE.}

\item{plot_background_fill}{Fill colour for the plot background. Default is
NULL.}
}
\value{
A \link{ggplot2} theme.
}
\description{
These are wrappers for \code{theme_paleta()} that use colours and fonts from the
Africa CDC visual identity guidelines.
}
\section{Colours}{

The Africa CDC theme is based on the colours from the \code{acdc_palettes}. The
primary palette consists of three colours: \code{acdc_palettes$acdc_primary}. The
secondary palette consists of five colours: \code{acdc_palettes$acdc_secondary}.
}

\section{Fonts}{

The Africa CDC theme uses one or up to two fonts from the four fonts
prescribed by the Africa CDC visual identity guidelines. These fonts (in
hierarchical order of preference) are \emph{Acumin Pro}, \emph{Brandon Text},
\emph{Calibri}, and/or \emph{Arial}. Any or all of these fonts should be available in
the user's system for them to be used in the theme. If none of these fonts
are available in the user's system, a freely downloadable alternative called
\emph{Noto Sans} is the default fallback font and can be obtained from
\href{https://fonts.google.com/}{Google Fonts}.
}

\examples{
\dontrun{
  ggplot(
    data = mtcars,
    mapping = aes(
      x = factor(vs, levels = c(0, 1), labels = c("v-shaped", "straight")),
      fill = factor(cyl))
  ) +
  geom_bar() +
  scale_fill_manual(
    name = "Cylinders",
    values = acdc_palettes$acdc_secondary
  ) +
  labs(
    title = "Engine shape by number of cylinders",
    subtitle = "An example plot for this package",
    x = "Engine Shape",
    y = "Counts"
   ) +
   theme_acdc_light()
}

}
