% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_unicef.R
\name{theme_unicef}
\alias{theme_unicef}
\title{A \link{ggplot2} theme using UNICEF fonts, colours, and palettes}
\usage{
theme_unicef(
  base_family = set_unicef_font(),
  plot_title_family = base_family,
  plot_title_colour = unicef_black,
  subtitle_family = base_family,
  subtitle_colour = unicef_cool_grey,
  caption_colour = unicef_cool_grey,
  axis_title_colour = unicef_cool_grey,
  legend_title_colour = unicef_cool_grey,
  legend_text_colour = unicef_cool_grey,
  grid_col = unicef_warm_grey,
  grid = TRUE,
  axis_col = unicef_warm_grey,
  axis = FALSE,
  ticks = FALSE
)
}
\arguments{
\item{base_family}{Base font family using UNICEF fonts. Default is set
by what UNICEF font is available in the system via \code{set_unicef_font()}.
If none of the UNICEF fonts are available, the default becomes
\emph{Noto Sans}.}

\item{plot_title_family}{Font family to use for the plot title. Default is
\code{base_family}.}

\item{plot_title_colour}{Colour of the plot title text. Default
is \code{unicef_blue}.}

\item{subtitle_family}{Font family to use for the plot subtitle. Default is
\code{base_family}.}

\item{subtitle_colour}{Colour of the subtitle text. Default is
\code{unicef_cool_grey}.}

\item{caption_colour}{Colour of the caption text. Default is
\code{unicef_cool_grey}.}

\item{axis_title_colour}{Colour of the axis title text. Default is
\code{unicef_cool_grey}.}

\item{legend_title_colour}{Colour of the legend title text. Default is NULL.}

\item{legend_text_colour}{Colour of the legend text. Default is NULL.}

\item{grid_col}{Grid colour. Default to \code{unicef_warm_grey}.}

\item{grid}{Panel grid. Either \code{TRUE}, \code{FALSE}, or a combination of
\code{X} (major x grid), \code{x} (minor x grid), \code{Y} (major y grid), and/or
\code{y} (minor y grid). Default is TRUE.}

\item{axis_col}{Axis colours. Default to \code{unicef_warm_grey}.}

\item{axis}{Add x or y axes? \code{TRUE}, \code{FALSE}, "\code{xy}". Default is FALSE.}

\item{ticks}{Logical. Should ticks be added? Default is FALSE.}
}
\value{
A \link{ggplot2} theme.
}
\description{
These are wrappers for \code{theme_paleta()} that use colours and fonts from the
UNICEF visual identity guidelines.
}
\section{Colours}{

The UNICEF theme is based on the colours from the \code{unicef_palettes}. The
primary palette consists of one colour: \code{unicef_palettes$unicef_primary}.
The secondary palette consists of eleven colours:
\code{unicef_palettes$unicef_secondary}.
}

\section{Fonts}{

The UNICEF theme uses two fonts as prescribed by the UNICEF visual identity
guidelines. These fonts (in hierarchical order of preference) are
\emph{Univers LT Pro}, \emph{Arial}, \emph{Roboto}, and \emph{Aleo}. Any or all of these fonts
should be available in the user's system for them to be used in the theme.
If none of these fonts are available in the user's system, a freely
downloadable alternative called \emph{Noto Sans} is the default fallback font and
can be obtained from \href{https://fonts.google.com/}{Google Fonts}.
}

\examples{
\dontrun{
  ggplot(
    data = mtcars,
    mapping = aes(
      x = factor(vs, levels = c(0, 1), labels = c("v-shaped", "straight")),
      fill = factor(cyl))
  ) +
  geom_bar() +
  scale_fill_manual(
    name = "Cylinders",
    values = unicef_palettes$unicef_secondary
  ) +
  labs(
    title = "Engine shape by number of cylinders",
    subtitle = "An example plot for this package",
    x = "Engine Shape",
    y = "Counts"
   ) +
   theme_unicef()
}

}
