% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{tint_colour}
\alias{tint_colour}
\alias{tint_colours}
\title{Get tint of colours}
\usage{
tint_colour(hex, p)

tint_colours(hex, p)
}
\arguments{
\item{hex}{A character value or vector of character of values for hex code
of colour/s to tint.}

\item{p}{Range from 0 to 1 for proportion to tint the colour/s with.}
}
\value{
A character value or vector of character values of hex code/s
tinted to the desired proportion.
}
\description{
Get tint of colours
}
\examples{
tint_colour(acdc_green, p = 0.2)
tint_colours(acdc_palettes$acdc_palettes$acdc_secondary, p = 0.4)

}
