% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_unicef.R
\docType{data}
\name{unicef_blue}
\alias{unicef_blue}
\alias{unicef_green}
\alias{unicef_lime_green}
\alias{unicef_yellow}
\alias{unicef_orange}
\alias{unicef_bright_red}
\alias{unicef_dark_red}
\alias{unicef_purple}
\alias{unicef_warm_grey}
\alias{unicef_cool_grey}
\alias{unicef_black}
\alias{unicef_dark_blue}
\title{UNICEF colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
unicef_blue

unicef_green

unicef_lime_green

unicef_yellow

unicef_orange

unicef_bright_red

unicef_dark_red

unicef_purple

unicef_warm_grey

unicef_cool_grey

unicef_black

unicef_dark_blue
}
\description{
UNICEF colours
}
\examples{
unicef_blue
unicef_green

}
\keyword{datasets}
