% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\docType{data}
\name{wb_blue}
\alias{wb_blue}
\alias{wb_cyan}
\alias{wb_black}
\alias{wb_white}
\alias{wb_bright_orange}
\alias{wb_bright_yellow}
\alias{wb_bright_red}
\alias{wb_light_orange}
\alias{wb_bright_aqua}
\alias{wb_bright_green}
\alias{wb_bright_purple}
\alias{wb_light_aqua}
\alias{wb_dark_red}
\alias{wb_dark_orange}
\alias{wb_brown}
\alias{wb_dark_purple}
\alias{wb_dark_aqua}
\alias{wb_dark_green}
\title{World Bank colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
wb_blue

wb_cyan

wb_black

wb_white

wb_bright_orange

wb_bright_yellow

wb_bright_red

wb_light_orange

wb_bright_aqua

wb_bright_green

wb_bright_purple

wb_light_aqua

wb_dark_red

wb_dark_orange

wb_brown

wb_dark_purple

wb_dark_aqua

wb_dark_green
}
\description{
World Bank colours
}
\examples{
wb_blue
wb_cyan

}
\keyword{datasets}
