% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\docType{data}
\name{wb_blue}
\alias{wb_blue}
\alias{wb_cyan}
\alias{wb_black}
\alias{wb_white}
\alias{wb_torrid}
\alias{wb_electric}
\alias{wb_rojo}
\alias{wb_carrot}
\alias{wb_turquoise}
\alias{wb_green}
\alias{wb_purple}
\alias{wb_teal}
\alias{wb_red}
\alias{wb_orange}
\alias{wb_brown}
\alias{wb_aubergine}
\alias{wb_beryl}
\alias{wb_emerald}
\title{World Bank colours}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
wb_blue

wb_cyan

wb_black

wb_white

wb_torrid

wb_electric

wb_rojo

wb_carrot

wb_turquoise

wb_green

wb_purple

wb_teal

wb_red

wb_orange

wb_brown

wb_aubergine

wb_beryl

wb_emerald
}
\description{
World Bank colours
}
\examples{
wb_blue
wb_cyan

}
\keyword{datasets}
