% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_wb.R
\docType{data}
\name{wb_fonts}
\alias{wb_fonts}
\alias{set_wb_font}
\title{World Bank fonts}
\format{
An object of class \code{list} of length 2.
}
\usage{
wb_fonts

set_wb_font(alt = paleta_fonts$paleta_noto)
}
\arguments{
\item{alt}{A character value for font family to use if all of the World Bank
fonts are not available in the system.}
}
\value{
A character value for font family to use as World Bank font.
}
\description{
The function will search the system for availability of any of the World Bank
fonts in heirarchical order starting with \emph{Andes}, and then \emph{Arial}. If
none of these are found in the system, the function will return \emph{Noto Sans}
by default or the user can set which font to use as alternative by specifying
\code{alt}.
}
\examples{
wb_fonts

set_wb_font()

}
\keyword{datasets}
